requirements.txt
aiohttp==3.8.4
beautifulsoup4==4.12.2
cryptography==40.0.2
Flask==2.3.2
python-dotenv==1.0.0
openai==0.27.8
requests==2.31.0
socksio==1.0.0